refactor(navigation): implement MainScreen with improved nav handling
