+++ app/src/main/java/com/example/shmr_finance_app_android/ui/navigation/BottomNavigationBar.kt
@@ -9,24 +9,19 @@ import androidx.compose.material3.NavigationBarItem
 import androidx.compose.material3.NavigationBarItemDefaults
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.navigation.NavController
-import androidx.navigation.NavGraph.Companion.findStartDestination
-import androidx.navigation.compose.currentBackStackEntryAsState
 import com.example.shmr_finance_app_android.R
 import com.example.shmr_finance_app_android.navigation.RootScreen
 
 @Composable
 fun BottomNavigationBar(
-    navController: NavController,
+    currentDestination: String?,
+    onNavigate: (String) -> Unit,
     screens: List<RootScreen>
 ) {
-    val navBackStackEntry by navController.currentBackStackEntryAsState()
-    val currentDestination = navBackStackEntry?.destination?.route
 
     NavigationBar {
         screens.forEach { screen ->
@@ -44,15 +39,7 @@ fun BottomNavigationBar(
                         contentDescription = stringResource(screen.label)
                     )
                 },
-                onClick = {
-                    navController.navigate(screen.route) {
-                        popUpTo(navController.graph.findStartDestination().id) {
-                            saveState = true
-                        }
-                        launchSingleTop = true
-                        restoreState = true
-                    }
-                },
+                onClick = { onNavigate(screen.route) },
                 colors = NavigationBarItemDefaults.colors(
                     indicatorColor = MaterialTheme.colorScheme.onTertiaryContainer,
                     selectedIconColor = MaterialTheme.colorScheme.tertiary,
